requirements.txt
blis==0.7.11
catalogue==2.0.10
certifi==2024.2.2
charset-normalizer==3.3.2
click==8.1.7
confection==0.1.4
contourpy==1.2.0
cycler==0.12.1
cymem==2.0.8
datasketch==1.6.4
emoji==1.4.2
filelock==3.13.1
fonttools==4.47.2
fsspec==2024.3.0
huggingface-hub==0.21.4
idna==3.6
Jinja2==3.1.3
joblib==1.3.2
kiwisolver==1.4.5
langcodes==3.3.0
MarkupSafe==2.1.5
matplotlib==3.7.4
mpmath==1.3.0
murmurhash==1.0.10
networkx==3.2.1
nltk==3.8.1
numpy==1.26.3
nvidia-cublas-cu12==12.1.3.1
nvidia-cuda-cupti-cu12==12.1.105
nvidia-cuda-nvrtc-cu12==12.1.105
nvidia-cuda-runtime-cu12==12.1.105
nvidia-cudnn-cu12==8.9.2.26
nvidia-cufft-cu12==11.0.2.54
nvidia-curand-cu12==10.3.2.106
nvidia-cusolver-cu12==11.4.5.107
nvidia-cusparse-cu12==12.1.0.106
nvidia-nccl-cu12==2.18.1
nvidia-nvjitlink-cu12==12.4.99
nvidia-nvtx-cu12==12.1.105
packaging==23.2
pandas==1.5.3
pathlib_abc==0.1.1
pathy==0.11.0
pillow==10.2.0
preshed==3.0.9
pycountry==23.12.11
pydantic==1.10.14
pyparsing==3.1.1
python-dateutil==2.8.2
pytz==2023.3.post1
PyYAML==6.0.1
regex==2023.12.25
requests==2.31.0
scikit-learn==1.3.0
scipy==1.11.4
seaborn==0.13.1
six==1.16.0
smart-open==6.4.0
spacy==3.5.3
spacy-legacy==3.0.12
spacy-loggers==1.0.5
srsly==2.4.8
sympy==1.12
thinc==8.1.12
threadpoolctl==3.2.0
tokenizers==0.13.3
torch==2.1.0
tqdm==4.66.1
transformers==4.27.1
triton==2.1.0
tweet-preprocessor==0.6.0
typer==0.7.0
typing_extensions==4.10.0
tzdata==2023.4
Unidecode==1.3.0
urllib3==2.2.1
wasabi==1.1.2
wordcloud==1.9.2
